Define algorithm providers in terms of plugins.
