+++ pkg/scheduler/algorithm/priorities/metadata_test.go
@@ -155,7 +155,6 @@ func TestPriorityMetadata(t *testing.T) {
 		informerFactory.Core().V1().ReplicationControllers().Lister(),
 		informerFactory.Apps().V1().ReplicaSets().Lister(),
 		informerFactory.Apps().V1().StatefulSets().Lister(),
-		1,
 	)
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
