+++ pkg/scheduler/core/generic_scheduler_test.go
@@ -1117,7 +1117,6 @@ func TestZeroRequest(t *testing.T) {
 				informerFactory.Core().V1().ReplicationControllers().Lister(),
 				informerFactory.Apps().V1().ReplicaSets().Lister(),
 				informerFactory.Apps().V1().StatefulSets().Lister(),
-				1,
 			)
 
 			metadata := metadataProducer(test.pod, test.nodes, snapshot)
