+++ pkg/scheduler/framework/plugins/defaultpodtopologyspread/default_pod_topology_spread_test.go
@@ -357,7 +357,6 @@ func TestDefaultPodTopologySpreadScore(t *testing.T) {
 				fakelisters.ControllerLister(test.rcs),
 				fakelisters.ReplicaSetLister(test.rss),
 				fakelisters.StatefulSetLister(test.sss),
-				1,
 			)
 			metaData := metaDataProducer(test.pod, nodes, snapshot)
 
@@ -615,7 +614,6 @@ func TestZoneSelectorSpreadPriority(t *testing.T) {
 				fakelisters.ControllerLister(test.rcs),
 				fakelisters.ReplicaSetLister(test.rss),
 				fakelisters.StatefulSetLister(test.sss),
-				1,
 			)
 			metaData := metaDataProducer(test.pod, nodes, snapshot)
