+++ pkg/scheduler/framework/plugins/serviceaffinity/service_affinity_test.go
@@ -403,8 +403,7 @@ func TestServiceAffinityScore(t *testing.T) {
 				fakelisters.ServiceLister(test.services),
 				fakelisters.ControllerLister(rcs),
 				fakelisters.ReplicaSetLister(rss),
-				fakelisters.StatefulSetLister(sss),
-				1)
+				fakelisters.StatefulSetLister(sss))
 			metaData := metaDataProducer(test.pod, nodes, snapshot)
 			state := framework.NewCycleState()
 			state.Write(migration.PrioritiesStateKey, &migration.PrioritiesStateData{Reference: metaData})
